Fix spline::synchronize() sometimes not converging
The dichotomy for reducing phase D velocity would sometimes stop before optimum
because of a too small threshold. This could lead to a longer duration than
expected, and in particular a non-zero final jerk.

To fix this, consistently use EPSILON (1e-10) among all dichotomies in the
slpine class, instead of a random mix between EPI4 (1e-4) and EPSI4/100
(1e-6).